+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - The config file default path is now shown in the help for sensu-backend start
 and sensu-agent start.
 - Keepalives can now be published via the HTTP API.
+- Token substitution templates can now express escape-quoted strings.
 
 ## [5.19.0] - 2020-03-26